Removed dtype

+++ main.py
@@ -4,7 +4,7 @@ import torch
 model_id = "mann-e/mann-e_4_rev-1-3"
 scheduler = EulerAncestralDiscreteScheduler.from_pretrained(model_id, subfolder="scheduler")
 
-pipe = StableDiffusionPipeline.from_pretrained(model_id, scheduler=scheduler, torch_dtype=torch.float16)
+pipe = StableDiffusionPipeline.from_pretrained(model_id, scheduler=scheduler)
 pipe = pipe.to("cpu")
 
 def dummy(images, **kwargs):
